+++ examples/test/with/test_with_badmulti_second.pas
@@ -1,0 +1,10 @@
+program TestWithBadMultiSecond;
+type
+    TPoint = record n: integer; end;
+var
+    a: TPoint;
+begin
+    with a, nothere do begin
+        n := 1;
+    end;
+end.